+++ src/components/player.component.js
@@ -59,7 +59,7 @@
  *  - [Boolean] **opl-veo-labs-icon** true to display the Veo-Labs icon (default to true)
  *  - [Boolean] **opl-time** true to display the actual time and duration (default to true)
  *  - [Boolean] **opl-chapters-tab** true to display chapters tab (default to true)
- *  - [Boolean] **opl-hide-tags-tab** true to hide tags tab (default to false)
+ *  - [Boolean] **opl-tags-tab** true to display tags tab (default to true)
  *  - [Boolean] **opl-full-viewport** true to display the player in full viewport (default to false)
  *  - [Boolean] **opl-disable-cut** true to disable cuts (default to false)
  *  - [String] **opl-language** Player language code (e.g. fr) (default to "en")
@@ -274,7 +274,7 @@
  *                  opl-auto-play="false"
  *                  opl-remember-position="false"
  *                  opl-chapters-tab="true"
- *                  opl-hide-tags-tab="false"
+ *                  opl-tags-tab="true"
  *                  opl-disable-cut="false"
  *      ></opl-player>
  *
@@ -300,7 +300,7 @@
       oplAutoPlay: '@?',
       oplRememberPosition: '@?',
       oplChaptersTab: '@?',
-      oplHideTagsTab: '@?',
+      oplTagsTab: '@?',
       oplDisableCut: '@?'
     }
   });