+++ src/components/player.component.spec.js
@@ -559,7 +559,7 @@ describe('OplPlayer', function() {
     assert.equal(tabsElement.attr('opl-no-tabs'), 'true', 'Unexpected tabs');
   });
 
-  it('should be able to hide tags using attribute "opl-hide-tags-tab"', function() {
+  it('should be able to hide tags using attribute "opl-tags-tab"', function() {
     scope.data.tags = [
       {
         value: 1000,
@@ -570,7 +570,7 @@ describe('OplPlayer', function() {
     var element = angular.element('<opl-player ' +
                                               'id="opl-player-test" ' +
                                               'opl-data="data" ' +
-                                              'opl-hide-tags-tab="true" ' +
+                                              'opl-tags-tab="false" ' +
                                   '></opl-player>');
     createComponent(element, 10000);