+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Player attribute "opl-mode" available values are now "split_1", "split_2", "split_50_50" and "split_25_75" instead of "media", "presentation", "both" and "both-presentation"
 - Player attribute "opl-mode-icon" has been renamed into "opl-template-icon"
 - Player attribute "opl-hide-chapters-tab" has been renamed into "opl-chapters-tab" and has been inverted (replace "false" by "true" and "true" by "false")
+- Player attribute "opl-hide-tags-tab" has been renamed into "opl-tags-tab" and has been inverted (replace "false" by "true" and "true" by "false")
 - Sources have been added to the settings menu, consequently the attribute "opl-media-sources-icon" has been removed
 - Data property "file.basePath" for tags and chapters has been renamed into "file.url"
 - Data property "file.originalname" for tags and chapters has been renamed into "file.originalName"